NP-1202 Extract role type
