+++ src/types/role.types.ts
@@ -1,0 +1,11 @@
+import { RoleName } from './user.types';
+
+interface Role {
+  rolename: RoleName;
+}
+
+export interface UserRoles {
+  institution: string;
+  roles: Role[];
+  username: string;
+}
